ban and unban, superadmin and users

+++ Controllers/AdminController.cs
@@ -3,25 +3,45 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using net_il_mio_fotoalbum.Models;
+using System.Security.Claims;
 
 namespace net_il_mio_fotoalbum.Controllers
 {
-    [Authorize(Roles = "ADMIN")]
+    [Authorize]
     public class AdminController : Controller
     {
         //Index
         public IActionResult Index(IndexFilter? data)
         {
             if (data == null) data = new IndexFilter();
-            using(PhotoContext ctx = new PhotoContext())
+            if (User.IsInRole("ADMIN"))
             {
-                if (data.Filter == null) data.Filter = "";
-                List<Photo> photos = ctx.Photos
-                    .Include(p => p.Tags)
-                    .Where(p => p.Title.Contains(data.Filter))
-                    .ToList();
-                data.Photos = photos;
-                return View(data);
+                using (PhotoContext ctx = new PhotoContext())
+                {
+                    if (data.Filter == null) data.Filter = "";
+                    List<Photo> photos = ctx.Photos
+                        .Include(p => p.Tags)
+                        .Where(p => p.Title.Contains(data.Filter))
+                        .ToList();
+                    data.Photos = photos;
+                    return View(data);
+                }
+
+            }
+            else
+            {
+                using (PhotoContext ctx = new PhotoContext())
+                {
+                    if (data.Filter == null) data.Filter = "";
+                    List<Photo> photos = ctx.Photos
+                        .Include(p => p.Tags)
+                        .Where(p => p.Title.Contains(data.Filter))
+                        .Where(p => p.UserId == User.FindFirst(ClaimTypes.NameIdentifier).Value)
+                        .ToList();
+                    data.Photos = photos;
+                    return View(data);
+                }
+
             }
         }
 
@@ -39,6 +59,7 @@ namespace net_il_mio_fotoalbum.Controllers
         [HttpGet]
         public IActionResult Create()
         { 
+
             PhotoFormModel model = new PhotoFormModel();
             List<Tag> listTags = new List<Tag>();
 			using (PhotoContext ctx = new PhotoContext())
@@ -49,7 +70,9 @@ namespace net_il_mio_fotoalbum.Controllers
                 }
             }
             model.Tags = listTags;
+
             model.Photo = new Photo();
+
             return View(model);
         }
 
@@ -58,11 +81,10 @@ namespace net_il_mio_fotoalbum.Controllers
         public IActionResult Create(PhotoFormModel data)
         {
             data.Tags = new List<Tag>();
-
             //Model validation
             if(!ModelState.IsValid)
             {
-                //var errors = ModelState.Values.SelectMany(x => x.Errors);
+                var errors = ModelState.Values.SelectMany(x => x.Errors);
                 using(PhotoContext ctx = new PhotoContext())
                 {
                     data.Tags = ctx.Tags.ToList();
@@ -84,7 +106,7 @@ namespace net_il_mio_fotoalbum.Controllers
                     }
 
                 }
-                
+                data.Photo.UserId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
 
                 ctx.Photos.Add(data.Photo);
                 ctx.SaveChanges();
@@ -100,7 +122,10 @@ namespace net_il_mio_fotoalbum.Controllers
             using(PhotoContext ctx = new PhotoContext())
             {
                 form.Photo = ctx.Photos.Where(p => p.Id == id).Include(p => p.Tags).FirstOrDefault();
-
+                if(form.Photo.UserId != User.FindFirst(ClaimTypes.NameIdentifier).Value)
+                {
+                    return NotFound();
+                }
 				if (form.Photo != null)
 				{
                     form.Tags = ctx.Tags.ToList();  
@@ -156,7 +181,7 @@ namespace net_il_mio_fotoalbum.Controllers
             using(PhotoContext ctx = new PhotoContext())
             {
                 Photo record = ctx.Photos.Where(p => p.Id == id).FirstOrDefault();
-                if(record == null) return NotFound();
+                if(record == null || record.UserId != User.FindFirst(ClaimTypes.NameIdentifier).Value) return NotFound();
 
                 ctx.Photos.Remove(record);
                 ctx.SaveChanges();
@@ -212,6 +237,15 @@ namespace net_il_mio_fotoalbum.Controllers
 			return Redirect("Index");
 		}
 
-
+        public IActionResult Moderate(int id)
+        {
+            using(PhotoContext ctx = new PhotoContext())
+            {
+                Photo record = ctx.Photos.Where(r => r.Id == id).FirstOrDefault();
+                record.Moderate = !record.Moderate;
+                ctx.SaveChanges();
+            }
+            return Redirect("/Admin");
+        }
 	}
 }
